refactor(relaxtime): split gap transport scan script shell
