+++ docs/dev/backlog/2026-05-04_sysimage产品化与package-app-service演进任务单.md
@@ -82,6 +82,7 @@
   - 2026-05-04：已拉起 `docs/dev/active/2026-05-04_sysimage产品化_PhaseC_package化_C1薄封装盘点任务单.md` 并完成首轮盘点；`run_gap_transport_scan.jl` 被识别为首要收敛对象。
 - [ ] C2 把稳定 CLI 执行路径进一步收敛到 `src/models` / `src/simulation` 暴露的稳定 API
   - 验收：脚本更多承担参数解析与入口职责
+  - 2026-05-04：已拉起 `docs/dev/active/2026-05-04_sysimage产品化_PhaseC_package化_C2首批边界收敛任务单.md` 并完成首批收敛；`run_gap_transport_scan.jl` 的 CLI 与脚本 IO 层已抽离为 helper，但核心求解执行体仍保留在脚本中。
 - [ ] C3 扩展 precompile capability 与 workload，使其绑定稳定 API 而非脚本偶然路径
   - 验收：precompile registry 与 stable API 对齐
