+++ scripts/relaxtime/run_gap_transport_scan.jl
@@ -56,299 +56,18 @@ const MODULE_DEFAULT_PHI_NODES = RT_ASR.DEFAULT_PHI_NODES       # 8
 const MODULE_DEFAULT_SIGMA_GRID_N = RT_ASR.DEFAULT_SIGMA_GRID_N # 60
 const PHASE_BOUNDARY_XI_CACHE = Ref{Union{Nothing, Vector{Float64}}}(nothing)
 const PHASE_CROSSOVER_XI_CACHE = Ref{Union{Nothing, Vector{Float64}}}(nothing)
-struct ScanOptions
-    output::String
-    channel_diagnostics_output::Union{Nothing, String}
-    failed_points_output::Union{Nothing, String}
-    provenance_dir::Union{Nothing, String}
-    xi_values::Vector{Float64}
-    tmin_mev::Float64
-    tmax_mev::Float64
-    tstep_mev::Float64
-    mubmin_mev::Float64
-    mubmax_mev::Float64
-    mubstep_mev::Float64
-    overwrite::Bool
-    resume::Bool
-    compute_bulk::Bool
-    p_num::Int
-    t_num::Int
-    max_iter::Int
-    # tau / cross-section settings
-    tau_p_nodes::Int
-    tau_angle_nodes::Int
-    tau_phi_nodes::Int
-    tau_n_sigma_points::Int
-    tau_threshold_subtraction::Bool
-    tau_asym_window::Float64
-    tau_asym_fit_min_points::Int
-    tau_asym_extra_points::Int
-    tau_interpolation_mode::Symbol
-    sigma_grid_n::Int
-    integration_mode::Symbol
-    gc_every_n::Int
-    # transport settings
-    tr_p_nodes::Int
-    tr_p_max_fm::Float64
-end
+include(joinpath(PROJECT_ROOT, "scripts", "relaxtime", "gap_transport_scan_cli.jl"))
+include(joinpath(PROJECT_ROOT, "scripts", "relaxtime", "gap_transport_scan_io.jl"))
 
-function print_usage()
-    println("Usage: julia --project=. scripts/relaxtime/run_gap_transport_scan.jl [options]\n")
-    println("Options:")
-    println("  --output <path>             输出 CSV (default data/outputs/results/relaxtime/scan/gap_transport_scan.csv)")
-    println("  --channel-diagnostics-output <path>  可选：输出每点每通道的 τ^-1 贡献明细 CSV")
-    println("  --failed-points-output <path>        可选：输出跳过失败点 sidecar CSV")
-    println("  --provenance-dir <path>              可选：写入 run_manifest/effective_config 的目录（默认 output 同目录）")
-    println("  --xi <value>                追加一个 ξ 值（可多次传入）")
-    println("  --xi-list v1,v2,...         用逗号分隔的 ξ 列表替换")
-    println("  --tmin/--tmax/--tstep <MeV> 温度范围与步长")
-    println("  --mubmin/--mubmax/--mubstep <MeV> 重子化学势 μ_B 范围与步长")
-    println("  --mumin/--mumax/--mustep <MeV> (兼容旧参数) 夸克化学势 μ_q 范围与步长")
-    println("  --overwrite                 覆盖输出文件")
-    println("  --no-resume                 禁用跳过逻辑，强制重算")
-    println("  --compute-bulk              计算体粘滞 ζ（很慢；默认关闭）")
-    println("  --p-num <int>               能隙/密度的动量节点数 (default 12)")
-    println("  --t-num <int>               能隙/密度的角度节点数 (default 6)")
-    println("  --max-iter <int>            NLsolve iterations 上限 (default 40)")
-    println("  --tau-p-nodes <int>         τ 平均散射率动量节点 (default $(MODULE_DEFAULT_P_NODES))")
-    println("  --tau-angle-nodes <int>     τ 平均散射率 cosθ 节点 (default $(MODULE_DEFAULT_ANGLE_NODES))")
-    println("  --tau-phi-nodes <int>       τ 平均散射率 φ 节点 (default $(MODULE_DEFAULT_PHI_NODES))")
-    println("  --tau-n-sigma <int>         σ(s) 的 t 积分点数 (default $(RT_TCS.DEFAULT_T_INTEGRAL_POINTS))")
-    println("  --tau-threshold-subtraction / --tau-no-threshold-subtraction  是否启用阈值减法缓存 (default true)")
-    println("  --tau-asym-window <float>   阈值减法拟合窗口 Δs (default 0.6)")
-    println("  --tau-asym-fit-min-points <int> 阈值减法最小拟合点数 (default 8)")
-    println("  --tau-asym-extra-points <int> 阈值减法补点数量 (default 10)")
-    println("  --tau-interpolation-mode <pchip|linear|direct|hybrid_threshold>  τ 中 σ(s) 取值模式 (default linear)")
-    println("  --sigma-grid-n <int>        σ(s) 预计算网格点数 (default $(MODULE_DEFAULT_SIGMA_GRID_N))")
-    println("  --mode <mode>               τ 积分模式: semi_infinite | finite_15 | finite_lambda (default semi_infinite)")
-    println("  --gc-every-n <int>          每 N 个点触发一次 GC (default 5; 0 表示关闭)")
-    println("  --tr-p-nodes <int>          输运积分动量节点 (default 24)")
-    println("  --tr-p-max <fm^-1>          输运积分 p 上限 (default 8.0)")
-    println("  -h, --help                  显示帮助")
-end
+using .GapTransportScanCLI: ScanOptions, parse_args, print_usage
+using .GapTransportScanIO: write_channel_diagnostics_header_if_needed,
+    write_failed_points_header_if_needed,
+    write_failed_point_row!,
+    ensure_parent_dir,
+    current_git_commit,
+    ensure_output_header_compatible,
+    write_header_if_needed
 
-function parse_args(args::Vector{String})
-    opts = Dict{Symbol,Any}(
-        :output => joinpath("data", "outputs", "results", "relaxtime", "scan", "gap_transport_scan.csv"),
-        :channel_diagnostics_output => nothing,
-        :failed_points_output => nothing,
-        :provenance_dir => nothing,
-        :xi_values => Float64[0.0],
-        :tmin => 50.0,
-        :tmax => 200.0,
-        :tstep => 10.0,
-        :mubmin => 0.0,
-        :mubmax => 1200.0,
-        :mubstep => 60.0,
-        :overwrite => false,
-        :resume => true,
-        :compute_bulk => false,
-        :p_num => 12,
-        :t_num => 6,
-        :max_iter => 40,
-        :tau_p_nodes => MODULE_DEFAULT_P_NODES,
-        :tau_angle_nodes => MODULE_DEFAULT_ANGLE_NODES,
-        :tau_phi_nodes => MODULE_DEFAULT_PHI_NODES,
-        :tau_n_sigma => RT_TCS.DEFAULT_T_INTEGRAL_POINTS,
-        :tau_threshold_subtraction => true,
-        :tau_asym_window => 0.6,
-        :tau_asym_fit_min_points => 8,
-        :tau_asym_extra_points => 10,
-        :tau_interpolation_mode => :linear,
-        :sigma_grid_n => MODULE_DEFAULT_SIGMA_GRID_N,
-        :integration_mode => :semi_infinite,
-        :gc_every_n => 5,
-        :tr_p_nodes => 24,
-        :tr_p_max => 8.0,
-    )
-
-    i = 1
-    while i <= length(args)
-        arg = args[i]
-        function require_value()
-            i == length(args) && error("missing value for $arg")
-            val = args[i + 1]
-            i += 1
-            return val
-        end
-        if arg == "--output"
-            opts[:output] = require_value()
-        elseif arg == "--channel-diagnostics-output"
-            opts[:channel_diagnostics_output] = require_value()
-        elseif arg == "--failed-points-output"
-            opts[:failed_points_output] = require_value()
-        elseif arg == "--provenance-dir"
-            opts[:provenance_dir] = require_value()
-        elseif arg == "--xi"
-            val = parse(Float64, require_value())
-            if opts[:xi_values] == Float64[0.0]
-                opts[:xi_values] = Float64[]
-            end
-            push!(opts[:xi_values], val)
-        elseif arg == "--xi-list"
-            raw = split(require_value(), ',')
-            vals = Float64[parse(Float64, strip(v)) for v in raw if !isempty(strip(v))]
-            opts[:xi_values] = vals
-        elseif arg == "--tmin"
-            opts[:tmin] = parse(Float64, require_value())
-        elseif arg == "--tmax"
-            opts[:tmax] = parse(Float64, require_value())
-        elseif arg == "--tstep"
-            opts[:tstep] = parse(Float64, require_value())
-        elseif arg == "--mubmin"
-            opts[:mubmin] = parse(Float64, require_value())
-        elseif arg == "--mubmax"
-            opts[:mubmax] = parse(Float64, require_value())
-        elseif arg == "--mubstep"
-            opts[:mubstep] = parse(Float64, require_value())
-        # 兼容旧参数：把 μ_q 转成 μ_B
-        elseif arg == "--mumin"
-            opts[:mubmin] = 3.0 * parse(Float64, require_value())
-        elseif arg == "--mumax"
-            opts[:mubmax] = 3.0 * parse(Float64, require_value())
-        elseif arg == "--mustep"
-            opts[:mubstep] = 3.0 * parse(Float64, require_value())
-        elseif arg == "--overwrite"
-            opts[:overwrite] = true
-        elseif arg == "--no-resume"
-            opts[:resume] = false
-        elseif arg == "--compute-bulk"
-            opts[:compute_bulk] = true
-        elseif arg == "--p-num"
-            opts[:p_num] = parse(Int, require_value())
-        elseif arg == "--t-num"
-            opts[:t_num] = parse(Int, require_value())
-        elseif arg == "--max-iter"
-            opts[:max_iter] = parse(Int, require_value())
-        elseif arg == "--tau-p-nodes"
-            opts[:tau_p_nodes] = parse(Int, require_value())
-        elseif arg == "--tau-angle-nodes"
-            opts[:tau_angle_nodes] = parse(Int, require_value())
-        elseif arg == "--tau-phi-nodes"
-            opts[:tau_phi_nodes] = parse(Int, require_value())
-        elseif arg == "--tau-n-sigma"
-            opts[:tau_n_sigma] = parse(Int, require_value())
-        elseif arg == "--tau-threshold-subtraction"
-            opts[:tau_threshold_subtraction] = true
-        elseif arg == "--tau-no-threshold-subtraction"
-            opts[:tau_threshold_subtraction] = false
-        elseif arg == "--tau-asym-window"
-            opts[:tau_asym_window] = parse(Float64, require_value())
-        elseif arg == "--tau-asym-fit-min-points"
-            opts[:tau_asym_fit_min_points] = parse(Int, require_value())
-        elseif arg == "--tau-asym-extra-points"
-            opts[:tau_asym_extra_points] = parse(Int, require_value())
-        elseif arg == "--tau-interpolation-mode"
-            mode = Symbol(require_value())
-            mode in (:pchip, :linear, :direct, :hybrid_threshold) || error("unknown tau interpolation mode: $(mode) (expected: pchip|linear|direct|hybrid_threshold)")
-            opts[:tau_interpolation_mode] = mode
-        elseif arg == "--sigma-grid-n"
-            opts[:sigma_grid_n] = parse(Int, require_value())
-        elseif arg == "--mode"
-            mode_str = require_value()
-            mode_sym = Symbol(mode_str)
-            mode_sym in (:semi_infinite, :finite_15, :finite_lambda) || error("unknown mode: $mode_str (expected: semi_infinite, finite_15, finite_lambda)")
-            opts[:integration_mode] = mode_sym
-        elseif arg == "--gc-every-n"
-            opts[:gc_every_n] = parse(Int, require_value())
-        elseif arg == "--tr-p-nodes"
-            opts[:tr_p_nodes] = parse(Int, require_value())
-        elseif arg == "--tr-p-max"
-            opts[:tr_p_max] = parse(Float64, require_value())
-        elseif arg in ("-h", "--help")
-            print_usage(); exit(0)
-        else
-            error("unknown option: $arg")
-        end
-        i += 1
-    end
-
-    xi_vals = opts[:xi_values]
-    isempty(xi_vals) && (xi_vals = Float64[0.0])
-    xi_vals = unique(sort(Float64.(xi_vals)))
-
-    tstep = Float64(opts[:tstep]); tstep > 0 || error("tstep must be positive")
-    mubstep = Float64(opts[:mubstep]); mubstep > 0 || error("mubstep must be positive")
-
-    return ScanOptions(
-        String(opts[:output]),
-        isnothing(opts[:channel_diagnostics_output]) ? nothing : String(opts[:channel_diagnostics_output]),
-        isnothing(opts[:failed_points_output]) ? nothing : String(opts[:failed_points_output]),
-        isnothing(opts[:provenance_dir]) ? nothing : String(opts[:provenance_dir]),
-        xi_vals,
-        Float64(opts[:tmin]),
-        Float64(opts[:tmax]),
-        tstep,
-        Float64(opts[:mubmin]),
-        Float64(opts[:mubmax]),
-        mubstep,
-        Bool(opts[:overwrite]),
-        Bool(opts[:resume]),
-        Bool(opts[:compute_bulk]),
-        Int(opts[:p_num]),
-        Int(opts[:t_num]),
-        Int(opts[:max_iter]),
-        Int(opts[:tau_p_nodes]),
-        Int(opts[:tau_angle_nodes]),
-        Int(opts[:tau_phi_nodes]),
-        Int(opts[:tau_n_sigma]),
-        Bool(opts[:tau_threshold_subtraction]),
-        Float64(opts[:tau_asym_window]),
-        Int(opts[:tau_asym_fit_min_points]),
-        Int(opts[:tau_asym_extra_points]),
-        Symbol(opts[:tau_interpolation_mode]),
-        Int(opts[:sigma_grid_n]),
-        Symbol(opts[:integration_mode]),
-        Int(opts[:gc_every_n]),
-        Int(opts[:tr_p_nodes]),
-        Float64(opts[:tr_p_max]),
-    )
-end
-
-function write_channel_diagnostics_header_if_needed(io)
-    header = join([
-        "T_MeV", "muq_MeV", "muB_MeV", "xi",
-        "species", "channel", "density_key", "multiplicity",
-        "density", "rate", "contribution", "total", "tau_inv_species",
-        "equilibrium_backend", "phase_curr", "phase_structure",
-    ], ',')
-    println(io, header)
-end
-
-@inline function _csv_quote(text::AbstractString)
-    return "\"" * replace(text, "\"" => "\"\"") * "\""
-end
-
-function write_failed_points_header_if_needed(io)
-    header = join([
-        "T_MeV", "muB_MeV", "xi",
-        "seed_source", "phase_prev", "phase_curr_hint",
-        "error_type", "error_message", "timestamp",
-    ], ',')
-    println(io, header)
-end
-
-function write_failed_point_row!(io, T_mev::Float64, muB_mev::Float64, xi::Float64, diag, err)
-    seed_source = hasproperty(diag, :seed_source) ? string(getproperty(diag, :seed_source)) : "unknown"
-    phase_prev = hasproperty(diag, :phase_prev) ? string(getproperty(diag, :phase_prev)) : "unknown"
-    phase_curr_hint = hasproperty(diag, :phase_curr) ? string(getproperty(diag, :phase_curr)) : "unknown"
-    error_type = string(typeof(err))
-    error_message = sprint(showerror, err)
-    timestamp = Dates.format(Dates.now(), dateformat"yyyy-mm-ddTHH:MM:SS")
-    row = join([
-        string(T_mev),
-        string(muB_mev),
-        string(xi),
-        _csv_quote(seed_source),
-        _csv_quote(phase_prev),
-        _csv_quote(phase_curr_hint),
-        _csv_quote(error_type),
-        _csv_quote(error_message),
-        _csv_quote(timestamp),
-    ], ',')
-    println(io, row)
-    flush(io)
-end
 
 @inline function _rate_with_alias(rates, key::Symbol)
     if hasproperty(rates, key)
@@ -483,81 +202,8 @@ function write_channel_diagnostics_rows!(io, T_mev::Float64, muq_mev::Float64, m
     flush(io)
 end
 
-function ensure_parent_dir(path::AbstractString)
-    dir = dirname(path)
-    isdir(dir) || mkpath(dir)
-end
-
-function current_git_commit()
-    try
-        return readchomp(`git -C $(PROJECT_ROOT) rev-parse HEAD`)
-    catch
-        return "unknown"
-    end
-end
-
 function read_existing_keys(path::AbstractString)
     return ScanCSV.read_existing_keys(path, ["T_MeV", "muB_MeV", "xi"])
-end
-
-function ensure_output_header_compatible(path::AbstractString)
-    isfile(path) || return
-    header = nothing
-    open(path, "r") do io
-        for line in eachline(io)
-            s = strip(line)
-            (isempty(s) || startswith(s, "#")) && continue
-            header = s
-            break
-        end
-    end
-    header === nothing && return
-    required = (
-        "omega_fm4inv",
-        "P_fm4inv",
-        "epsilon_fm4inv",
-        "s_fm3inv",
-        "eps_minus_3P_over_T4",
-        "eta_over_s",
-        "zeta_over_s",
-        "sigma_over_T",
-        "sigma_over_T_over_eta_over_s",
-        "zeta_over_s_over_eta_over_s",
-        "quality_flag",
-        "quality_reason",
-        "quality_metric",
-        "run_id",
-        "equilibrium_backend",
-        "seed_source",
-        "phase_prev",
-        "phase_curr",
-        "phase_structure",
-        "phase_boundary_xi_used",
-    )
-    for c in required
-        occursin(c, header) || error("existing output CSV header is incompatible with current script (missing column: $c). Please rerun with --overwrite or choose a new --output path.")
-    end
-end
-
-function write_header_if_needed(io)
-    header = join([
-        "T_MeV", "muq_MeV", "muB_MeV", "xi",
-        "T_fm", "muq_fm",
-        "converged", "iterations", "residual_norm", "equilibrium_backend", "seed_source", "phase_prev", "phase_curr", "phase_structure", "phase_boundary_xi_used",
-        "Phi", "Phibar",
-        "m_u", "m_d", "m_s",
-        "rho_baryon", "rho_norm",
-        "omega_fm4inv", "P_fm4inv", "epsilon_fm4inv", "s_fm3inv",
-        "omega_MeV_fm3", "P_MeV_fm3", "epsilon_MeV_fm3",
-        "eps_minus_3P_over_T4",
-        "n_u", "n_d", "n_s", "n_ubar", "n_dbar", "n_sbar",
-        "tau_u", "tau_d", "tau_s", "tau_ubar", "tau_dbar", "tau_sbar",
-        "tauinv_u", "tauinv_d", "tauinv_s", "tauinv_ubar", "tauinv_dbar", "tauinv_sbar",
-        "eta", "sigma", "zeta", "eta_over_s", "zeta_over_s",
-        "sigma_over_T", "sigma_over_T_over_eta_over_s", "zeta_over_s_over_eta_over_s",
-        "quality_flag", "quality_reason", "quality_metric", "run_id",
-    ], ',')
-    println(io, header)
 end
 
 function csv_bool(x::Bool)
